Crate mcrt

Monte-Carlo radiative transfer simulation binary. Compute the radiative field for a given set of setup and light source.

Constants

BACKUP_TERM_WIDTH
Backup print width if the terminal width can not be determined.

Functions

gen_base_output
Generate the base output instance.

gen_detector_registers
Generate the detector registers.

initialisation
Initialise the input arguments.

load_parameters
Load the required files and form the input parameters.

main
Main program function.